requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
Flask-Cors==3.0.10
passlib==1.7.4
passlib[bcrypt]==1.7.4